backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
bandit
python-multipart
python-jose[cryptography]
passlib[bcrypt]
boto3
slowapi
